Update text_to_speech.py
updated the speak function , make it more sturctured , also added to play sound using afplay in macos and redirected linux or other to aplay , make sure the  display of audio  only work for jupyter notebooks not  in terminal. the name of audio file is directly used in code , standerized it.